requirements.txt
geopandas 
rasterio 
shapely 
numpy
fastapi
python-multipart
uvicorn
motor
pydantic
pydantic[email]
python-dotenv
jinja2
requests
